# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
paths:
...
  /api/pelapor/pelaporan:
    post:
      summary: 'Create Laporan'
      operationId: createLaporan
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Pelapor
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                foto_bukti:
                  type: string
                  format: binary
                  description: 'Must be an image. Must not be greater than 2048 kilobytes.'
                  nullable: false
                deskripsi_laporan:
                  type: string
                  description: ''
                  example: officiis
                  nullable: false
                latitude:
                  type: string
                  description: 'Must match the regex /^(\+|-)?(?:90(?:\.0{1,15})?|[0-9]{1,2}(?:\.[0-9]{1,15})?)$/.'
                  example: '-:90:.00000000000|7:.7281640)'
                  nullable: false
                longitude:
                  type: string
                  description: 'Must match the regex /^(\+|-)?(?:180(?:\.0{1,15})?|((1[0-7][0-9]|[0-9]{1,2})(\.[0-9]{1,15})?))$/.'
                  example: '|176))'
                  nullable: false
                no_hp:
                  type: string
                  description: ''
                  example: null
                  nullable: false
              required:
                - foto_bukti
                - deskripsi_laporan
                - latitude
                - longitude
...
  /api/petugas/wilayah:
    get:
      summary: 'Get Wilayah'
      operationId: getWilayah
      description: ''
      parameters:
        -
          in: query
          name: is_paginate
          description: 'default 0'
          example: false
          required: false
          schema:
            type: boolean
            description: 'default 0'
            example: false
            nullable: false
        -
          in: query
          name: per_page
          description: 'default 10'
          example: 19
          required: false
          schema:
            type: integer
            description: 'default 10'
            example: 19
            nullable: false
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Petugas
...
components:
  securitySchemes:
    default:
      type: apiKey
      name: Authorization
      in: header
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []